PankratovRoman/requirements.txt
lxml~=4.6.3
beautifulsoup4~=4.10.0
requests~=2.26.0
python-dateutil~=2.8.2